requirements.txt
fastapi
uvicorn
pydantic
supabase
python-dotenv
asyncpg
colorama
httpx==0.27.2;
openai==1.55.3;
mcp
mcp-server-time
pydantic-ai
python-multipart
nest_asyncio==1.6.0;